refactor: remove Persistence
