+++ deployer/src/error.rs
@@ -1,4 +1,4 @@
-use super::deployment::persistence::dal::DalError;
+use crate::dal::DalError;
 
 /// A wrapper to capture any error possible with this service
 #[derive(thiserror::Error, Debug)]
