+++ deployer/src/main.rs
@@ -1,17 +1,14 @@
 use clap::Parser;
 use shuttle_common::backends::tracing::setup_tracing;
 use shuttle_deployer::{
-    args::Args,
-    deployment::persistence::{dal::Sqlite, Persistence},
-    runtime_manager::RuntimeManager,
-    DeployerService, DeployerServiceConfig, DeployerServiceConfigBuilder,
+    args::Args, dal::Sqlite, runtime_manager::RuntimeManager, DeployerService,
+    DeployerServiceConfig, DeployerServiceConfigBuilder,
 };
 use tracing::{error, trace};
 
 #[tokio::main]
 async fn main() {
     let args = Args::parse();
-    let (persistence, _) = Persistence::from_dal(Sqlite::new(&args.state).await).await;
     setup_tracing(tracing_subscriber::registry(), "deployer");
 
     trace!(args = ?args, "parsed args");
@@ -28,7 +25,7 @@ async fn main() {
         .expect("to build the deployer service configuration");
 
     let runtime_manager = RuntimeManager::new();
-    let svc = DeployerService::new(runtime_manager, persistence, config).await;
+    let svc = DeployerService::new(runtime_manager, Sqlite::new(&args.state).await, config).await;
 
     match svc.start().await {
         Ok(_) => (),
